:recycle: refactor : 도메인별 구조로 리팩토링
