+++ src/features/cart/styles/CartList.css
@@ -110,7 +110,6 @@
 .delete-btn:hover{
     background-color: #b6cc8e;
 }
-/* ... 기존 코드 ... */
 
 
 @media screen and (max-width: 1200px) {
