+++ src/features/enrollment/styles/waitingView.css
