+++ src/features/lecture/components/LectureRow.jsx
